requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0.post2
Flask-Migrate==4.0.5
Pillow==10.0.0
python-slugify==8.0.1
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Flask-Caching==2.0.2
Flask-Limiter==3.3.1
Flask-Assets==2.0
markdown2==2.4.10
bleach==6.0.0
boto3==1.28.25
redis==5.0.0
celery==5.3.1
PyJWT==2.8.0
